feat: 파일구조 이동 , 테스트 코드 임시 추가
