+++ src/main/java/com/example/AOD/domain/entity/Domain.java
@@ -1,0 +1,5 @@
+package com.example.AOD.domain.entity;
+
+public enum Domain {
+    AV, GAME, WEBTOON, WEBNOVEL
+}
