+++ src/main/java/com/example/AOD/repo/AvContentRepository.java
@@ -1,0 +1,6 @@
+package com.example.AOD.repo;
+
+import com.example.AOD.domain.entity.AvContent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AvContentRepository extends JpaRepository<AvContent, Long> { }
