+++ src/main/java/com/example/AOD/repo/GameContentRepository.java
@@ -1,0 +1,6 @@
+package com.example.AOD.repo;
+
+import com.example.AOD.domain.entity.GameContent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GameContentRepository extends JpaRepository<GameContent, Long> {}
