+++ src/main/java/com/example/AOD/repo/WebnovelContentRepository.java
@@ -1,0 +1,6 @@
+package com.example.AOD.repo;
+
+
+import com.example.AOD.domain.entity.WebnovelContent;
+import org.springframework.data.jpa.repository.JpaRepository;
+public interface WebnovelContentRepository extends JpaRepository<WebnovelContent, Long> {}
