+++ src/main/java/com/example/AOD/repo/WebtoonContentRepository.java
@@ -1,0 +1,5 @@
+package com.example.AOD.repo;
+
+import com.example.AOD.domain.entity.WebtoonContent;
+import org.springframework.data.jpa.repository.JpaRepository;
+public interface WebtoonContentRepository extends JpaRepository<WebtoonContent, Long> {}
