+++ src/main/java/com/example/AOD/rules/NormalizerStep.java
@@ -1,0 +1,14 @@
+package com.example.AOD.rules;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class NormalizerStep {
+    private String type;
+    private List<String> fields;
+    // getters/setters
+}
