+++ src/main/java/com/example/AOD/rules/webnovel/naverseries.yml
@@ -1,0 +1,21 @@
+platformName: NaverSeries
+domain: WEBNOVEL
+schemaVersion: 1
+fieldMappings:
+  data.titleName: master_title
+  data.authors[0].name: domain.author
+  data.serviceStartDate: domain.started_at
+  data.genres: domain.genres
+  data.thumbnailUrl: poster_image_url
+  data.synopsis: synopsis
+
+  data.titleId: platform.attributes.title_id
+  data.weekday: platform.attributes.weekday
+  data.episodeCount: platform.attributes.episode_count
+normalizers:
+  - type: nfkc
+    fields: [master_title, domain.author]
+  - type: strip_parentheses
+    fields: [master_title]
+  - type: collapse_spaces
+    fields: [master_title]
