--- src/main/java/com/example/AOD/v2/domain/Content.java
@@ -1,48 +1,0 @@
-package com.example.AOD.v2.domain;
-
-import com.example.AOD.v2.domain.entity.Domain;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.Instant;
-
-@Entity
-@Setter
-@Getter
-@Table(name = "contents",
-        indexes = @Index(name = "idx_contents_lookup", columnList = "domain,masterTitle,releaseYear"))
-public class Content {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long contentId;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Domain domain;
-
-    @Column(nullable = false)
-    private String masterTitle;
-
-    private String originalTitle;
-    private Integer releaseYear;
-    private String posterImageUrl;
-
-    @Column(columnDefinition = "text")
-    private String synopsis;
-
-    private Instant createdAt;
-    private Instant updatedAt;
-
-    @PrePersist
-    public void prePersist() {
-        createdAt = Instant.now();
-        updatedAt = createdAt;
-    }
-    @PreUpdate
-    public void preUpdate() {
-        updatedAt = Instant.now();
-    }
-
-    // getters/setters ...
-}
