--- src/main/java/com/example/AOD/v2/domain/entity/Domain.java
@@ -1,5 +1,0 @@
-package com.example.AOD.v2.domain.entity;
-
-public enum Domain {
-    AV, GAME, WEBTOON, WEBNOVEL
-}
