--- src/main/java/com/example/AOD/v2/repo/AvContentRepository.java
@@ -1,6 +1,0 @@
-package com.example.AOD.v2.repo;
-
-import com.example.AOD.v2.domain.entity.AvContent;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AvContentRepository extends JpaRepository<AvContent, Long> { }
