--- src/main/java/com/example/AOD/v2/repo/GameContentRepository.java
@@ -1,6 +1,0 @@
-package com.example.AOD.v2.repo;
-
-import com.example.AOD.v2.domain.entity.GameContent;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GameContentRepository extends JpaRepository<GameContent, Long> {}
