--- src/main/java/com/example/AOD/v2/repo/WebnovelContentRepository.java
@@ -1,6 +1,0 @@
-package com.example.AOD.v2.repo;
-
-
-import com.example.AOD.v2.domain.entity.WebnovelContent;
-import org.springframework.data.jpa.repository.JpaRepository;
-public interface WebnovelContentRepository extends JpaRepository<WebnovelContent, Long> {}
