--- src/main/java/com/example/AOD/v2/repo/WebtoonContentRepository.java
@@ -1,5 +1,0 @@
-package com.example.AOD.v2.repo;
-
-import com.example.AOD.v2.domain.entity.WebtoonContent;
-import org.springframework.data.jpa.repository.JpaRepository;
-public interface WebtoonContentRepository extends JpaRepository<WebtoonContent, Long> {}
