--- src/main/java/com/example/AOD/v2/rules/NormalizerStep.java
@@ -1,14 +1,0 @@
-package com.example.AOD.v2.rules;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class NormalizerStep {
-    private String type;
-    private List<String> fields;
-    // getters/setters
-}
