--- src/main/java/com/example/AOD/v2/service/RuleLoader.java
@@ -1,20 +1,0 @@
-package com.example.AOD.v2.service;
-
-import com.example.AOD.v2.rules.MappingRule;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.stereotype.Component;
-import org.yaml.snakeyaml.Yaml;
-
-import java.io.InputStream;
-
-@Component
-public class RuleLoader {
-    public MappingRule load(String pathOnClasspath) {
-        try (InputStream in = new ClassPathResource(pathOnClasspath).getInputStream()) {
-            Yaml yaml = new Yaml();
-            return yaml.loadAs(in, MappingRule.class);
-        } catch (Exception e) {
-            throw new IllegalStateException("Failed to load rule: " + pathOnClasspath, e);
-        }
-    }
-}
